+++ gecko_client/pchain_types.go
@@ -1,0 +1,28 @@
+package gecko_client
+
+type Validator struct {
+	StartTime string `json:"startTime"`
+	EndTime string	`json:"endTime"`
+	StakeAmount string	`json:"stakeAmount"`
+	Id string	`json:"id"`
+}
+
+type ValidatorList struct {
+	Validators []Validator	`json:"validators"`
+}
+
+type GetValidatorsResponse struct {
+	JsonRpcVersion string	`json:"jsonrpc"`
+	Result ValidatorList	`json:"result"`
+	Id int	`json:"id"`
+}
+
+type BlockchainStatus struct {
+	Status string	`json:"status"`
+}
+
+type GetBlockchainStatusResponse struct {
+	JsonRpcVersion string	`json:"jsonrpc"`
+	Result BlockchainStatus	`json:"result"`
+	Id int	`json:"id"`
+}