+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # TBD
 * Fix tee suppressing exit code of the Docker image
+* Add parameters to `makeRpcRequest` on Gecko client
+* Add `GetBlockchainStatus` endpoint on Gecko client
 
 # 0.2.0
 * Updated code to use latest Kurtosis version that stops containers